work in progress
Former-commit-id: e27a6ed5064bcbc0aadff6166e524f5bcb35bcc9
Former-commit-id: b0967ffa0cde3671b9e54bb45b238d73c32ede21

+++ src/wid_dirlist.c
@@ -331,7 +331,7 @@ static uint8_t wid_dirlist_receive_input (widp w, const SDL_KEYSYM *key)
 
                         wid_set_cursor(w, len);
 
-                        best = strlen(wid_get_text(child));
+                        best = (int) strlen(wid_get_text(child));
                     }
 
                     n = tree_get_next(r, r->node, n);
